Split service interfaces and default implementations in separate files
