+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/services/location/DefaultUserLocationService.kt
@@ -1,4 +1,4 @@
-package org.noiseplanet.noisecapture.services
+package org.noiseplanet.noisecapture.services.location
 
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
@@ -10,22 +10,6 @@ import org.noiseplanet.noisecapture.log.Logger
 import org.noiseplanet.noisecapture.model.Location
 import kotlin.random.Random
 import kotlin.time.Duration.Companion.seconds
-
-/**
- * Get user location updates
- */
-interface UserLocationService {
-
-    /**
-     * Gets current user location, if known
-     */
-    fun getCurrentLocation(): Location?
-
-    /**
-     * Tracks user location updates
-     */
-    fun getLiveLocation(): Flow<Location>
-}
 
 
 /**
