+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/services/measurement/DefaultMeasurementService.kt
@@ -1,4 +1,4 @@
-package org.noiseplanet.noisecapture.services
+package org.noiseplanet.noisecapture.services.measurement
 
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
@@ -6,24 +6,9 @@ import org.koin.core.parameter.parametersOf
 import org.noiseplanet.noisecapture.log.Logger
 import org.noiseplanet.noisecapture.model.Measurement
 
-interface MeasurementService {
-
-    /**
-     * Stores the given measurement to the database
-     *
-     * @param measurement Measurement to store
-     */
-    fun storeMeasurement(measurement: Measurement)
-
-    /**
-     * Gets locally stored measurements
-     *
-     * @return Measurements found in database
-     */
-    fun getMeasurements(): List<Measurement>
-}
-
-
+/**
+ * Default [MeasurementService] implementation
+ */
 class DefaultMeasurementService : MeasurementService, KoinComponent {
 
     // - Constants
