+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/services/settings/DefaultUserSettingsService.kt
@@ -1,4 +1,4 @@
-package org.noiseplanet.noisecapture.services
+package org.noiseplanet.noisecapture.services.settings
 
 import com.russhwolf.settings.ExperimentalSettingsApi
 import com.russhwolf.settings.Settings
@@ -10,38 +10,6 @@ import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.map
 import kotlinx.serialization.ExperimentalSerializationApi
-
-
-/**
- * Read and write persistent user settings as key value pairs.
- */
-interface UserSettingsService {
-
-    /**
-     * Sets the value associated to the given key.
-     *
-     * @param key User settings [SettingsKey]
-     * @param value New value
-     */
-    fun <T> set(key: SettingsKey<T>, value: T?)
-
-    /**
-     * Gets the value associated to a given key, or [SettingsKey.defaultValue] if value is not set.
-     *
-     * @param key User settings [SettingsKey]
-     * @return Value or [SettingsKey.defaultValue] if not found
-     */
-    fun <T> get(key: SettingsKey<T>): T
-
-    /**
-     * Gets a flow of values associated to a given key, starting with [SettingsKey.defaultValue]
-     * if value is not set.
-     *
-     * @param key User settings [SettingsKey]
-     * @return Value or [SettingsKey.defaultValue] if not found
-     */
-    fun <T> getFlow(key: SettingsKey<T>): Flow<T>
-}
 
 
 /**
