+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/measurement/indicators/AcousticIndicatorsViewModel.kt
@@ -2,7 +2,7 @@ package org.noiseplanet.noisecapture.ui.features.measurement.indicators
 
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.Flow
-import org.noiseplanet.noisecapture.services.LiveAudioService
+import org.noiseplanet.noisecapture.services.liveaudio.LiveAudioService
 
 class AcousticIndicatorsViewModel(
     private val measurementService: LiveAudioService,
