+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/measurement/plot/spectrogram/SpectrogramBitmap.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.unit.IntSize
 import org.noiseplanet.noisecapture.audio.signal.window.SpectrumData
 import org.noiseplanet.noisecapture.model.SpectrogramScaleMode
-import org.noiseplanet.noisecapture.services.DefaultLiveAudioService.Companion.FFT_SIZE
+import org.noiseplanet.noisecapture.services.liveaudio.DefaultLiveAudioService.Companion.FFT_SIZE
 import org.noiseplanet.noisecapture.util.toComposeColor
 import org.noiseplanet.noisecapture.util.toImageBitmap
 import org.noiseplanet.noisecapture.util.toLittleEndianBytes
