+++ composeApp/src/commonMain/kotlin/org/noiseplanet/noisecapture/ui/features/measurement/plot/spectrogram/SpectrogramPlotViewModel.kt
@@ -14,7 +14,7 @@ import kotlinx.coroutines.withContext
 import org.noiseplanet.noisecapture.audio.ANDROID_GAIN
 import org.noiseplanet.noisecapture.log.Logger
 import org.noiseplanet.noisecapture.model.SpectrogramScaleMode
-import org.noiseplanet.noisecapture.services.LiveAudioService
+import org.noiseplanet.noisecapture.services.liveaudio.LiveAudioService
 
 class SpectrogramPlotViewModel(
     private val measurementsService: LiveAudioService,
